+++ app/system/modules/comment/app/views/index.js
@@ -1,0 +1,11 @@
+const Comment = {
+    el: '#app',
+    name: 'Comment',
+    data() {
+        return _.merge({
+
+        }, window.$data)
+    }
+}
+export default Comment;
+Vue.ready(Comment)