+++ app/system/modules/comment/views/admin/index.php
@@ -1,0 +1,6 @@
+<?php $view->script('system_comment' , 'system/comment:app/bundle/index.js' , 'vue') ?>
+<div id="app" v-cloak>
+    <div class="">
+
+    </div>
+</div>